+++ example/basic_usage.dart
@@ -2,9 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:file_type/file_type.dart';
+import 'package:file_type_plus/file_type_plus.dart';
 
-/// Basic usage examples for the file_type package.
+/// Basic usage examples for the file_type_plus package.
 void main() {
   print('=== Basic File Type Detection ===\n');